@@ -2,9 +2,11 @@
   "name": "craftyapi",
   "version": "1.0.1",
   "description": "A API wrapper for the Minecraft server controller, CraftyControl.",
-  "main": "src/index.js",
+  "main": "dist/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc",
+    "prestart": "npm run build",
+    "start": "node ."
   },
   "keywords": [
     "crafty",
@@ -18,7 +20,12 @@
   "author": "SpiderJockey02",
   "license": "ISC",
   "devDependencies": {
-    "eslint": "^8.4.1"
+    "@types/node": "^18.11.11",
+    "@typescript-eslint/eslint-plugin": "^5.46.0",
+    "@typescript-eslint/parser": "^5.46.0",
+    "eslint": "^8.29.0",
+    "tslint": "^6.1.3",
+    "typescript": "^4.9.4"
   },
   "dependencies": {
     "axios": "^0.24.0"
